在linux系统测试通过

+++ src/main/java/cn/wisesign/Install.kt
@@ -5,6 +5,7 @@ import cn.wisesign.utils.Tools
 import cn.wisesign.utils.getMainAppDirectPath
 import org.apache.commons.io.FileUtils
 import java.io.File
+import java.io.File.separator
 import java.io.FileInputStream
 import java.io.FileOutputStream
 import java.io.IOException
@@ -58,25 +59,25 @@ class Install : CommandProcessor() {
     fun copyJavaExeForImsWindows(){
         val imsexe = "${imsHome}jre/bin/ims.exe"
         val javaexe = "${imsHome}jre/bin/java.exe"
-        if(!File(imsexe).exists()){
+        if(!File(imsexe).exists() && File(javaexe).exists()){
             FileUtils.copyFile(File(javaexe),File(imsexe))
         }
     }
 
     fun applySystemConfigProperties(){
-        var system_config = Properties()
+        val system_config = Properties()
         REAL_SYSTEM_CONFIG_PATH = "${Tools.getMainAppDirectPath(mainAppName)}$SYSTEM_CONFIG_PATH"
         println(REAL_SYSTEM_CONFIG_PATH)
         try {
             val ins = FileInputStream(REAL_SYSTEM_CONFIG_PATH)
             system_config.load(ins)
-            var out = FileOutputStream(REAL_SYSTEM_CONFIG_PATH)
-            system_config.put(DOC_LIBRARY_BASE_PATH, "$imsHome$DOC_LIBRARY_DIRECTOR_NAME")
-            system_config.put(ATTACHMENT_ROOT_PATH, "$imsHome$ATTACHMENT_DIRECTOR_NAME")
-            system_config.put(PROCESS_EXPORT_EXCEL_PATH, "$imsHome$PROCESS_EXPORT_DIRECTOR_NAME")
-            system_config.put(INDEX_PATH, "$imsHome$INDEX_DIRECTOR_NAME")
-            system_config.put(IMS_HOME, "$imsHome$IMS_HOME_PATH")
-            system_config.put(REPORT_EXEURL, "$imsHome$REPORT_PATH")
+            val out = FileOutputStream(REAL_SYSTEM_CONFIG_PATH)
+            system_config.put(DOC_LIBRARY_BASE_PATH, "$imsHome$separator$DOC_LIBRARY_DIRECTOR_NAME")
+            system_config.put(ATTACHMENT_ROOT_PATH, "$imsHome$separator$ATTACHMENT_DIRECTOR_NAME")
+            system_config.put(PROCESS_EXPORT_EXCEL_PATH, "$imsHome$separator$PROCESS_EXPORT_DIRECTOR_NAME")
+            system_config.put(INDEX_PATH, "$imsHome$separator$INDEX_DIRECTOR_NAME")
+            system_config.put(IMS_HOME, "$imsHome$separator$IMS_HOME_PATH")
+            system_config.put(REPORT_EXEURL, "$imsHome$separator$REPORT_PATH")
             system_config.store(out, "")
         } catch (e: IOException) {
             println(e.message)
